requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ── Image Processing ──────────────────────────────────────────────────────────
Pillow>=9.5.0
opencv-python>=4.7.0

# ── Data & Math ───────────────────────────────────────────────────────────────
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0

# ── Grad-CAM ─────────────────────────────────────────────────────────────────
grad-cam>=1.4.8

# ── RAG / Embeddings ─────────────────────────────────────────────────────────
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
groq>=0.4.2               # Free-tier LLM API

# ── Backend (Phase 6 — optional) ─────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
